BC-2476 Move error types to error enum in error directory
